refactor(DonutChart): move native slot types to types.ts
Address review feedback by colocating CellSlot/DonutSlots with the shared DonutChart types.
