+++ packages/blade/src/components/Charts/DonutChart/types.ts
@@ -2,6 +2,10 @@ import type { PieProps as RechartsPieProps, CellProps } from 'recharts';
 import type {
   ChartsCategoricalColorToken,
   ChartSequentialColorToken,
+  Layout,
+  Align,
+  ChartLegendProps,
+  ChartTooltipProps,
 } from '../CommonChartComponents/types';
 import type { ColorTheme } from '../utils';
 import type { BoxProps } from '~components/Box';
@@ -66,4 +70,44 @@ type ChartDonutCellProps = CellProps & {
   color?: ChartsCategoricalColorToken | ChartSequentialColorToken;
 };
 
-export type { ChartDonutWrapperProps, ChartDonutCellProps, ChartDonutProps, ChartRadius, Content };
+type CellSlot = {
+  color?: ChartsCategoricalColorToken | ChartSequentialColorToken;
+};
+
+type LegendSlot = {
+  selectedDataKeys?: string[];
+  defaultSelectedDataKeys?: string[];
+  onSelectedDataKeysChange?: ChartLegendProps['onSelectedDataKeysChange'];
+  layout: Layout;
+  align: Align;
+};
+
+type DonutSlot = {
+  data: Record<string, unknown>[];
+  dataKey: string;
+  nameKey: ChartDonutProps['nameKey'];
+  cx: ChartDonutProps['cx'];
+  cy: ChartDonutProps['cy'];
+  radius: ChartRadius;
+  colorTheme: ColorTheme;
+  type: NonNullable<ChartDonutProps['type']>;
+  cells: CellSlot[];
+};
+
+type DonutSlots = {
+  donut?: DonutSlot;
+  hasLegend: boolean;
+  legend?: LegendSlot;
+  hasTooltip: boolean;
+  tooltipFormatter?: NonNullable<ChartTooltipProps['formatter']>;
+};
+
+export type {
+  ChartDonutWrapperProps,
+  ChartDonutCellProps,
+  ChartDonutProps,
+  ChartRadius,
+  Content,
+  CellSlot,
+  DonutSlots,
+};
